add poll ID option for update_gac_scores, and local server
