+++ apps/consensus-service/api/update_gac_scores.py
@@ -12,7 +12,7 @@ import pandas as pd
 import math
 import aiohttp
 import asyncio
-from webhook_utils import send_webhook
+from .webhook_utils import send_webhook
 
 # Set pandas option for future-proof behavior with downcasting
 pd.set_option('future.no_silent_downcasting', True)
@@ -88,6 +88,69 @@ class handler(BaseHTTPRequestHandler):
             self.end_headers()
             self.wfile.write(b'Error updating GAC scores')
         return
+    
+    def do_POST(self):
+        # Check for API key if configured
+        api_key = os.environ.get('API_KEY')
+        if api_key:
+            request_key = self.headers.get('X-API-Key')
+            if not request_key or request_key != api_key:
+                self.send_response(401)
+                self.send_header('Content-type', 'application/json')
+                self.end_headers()
+                self.wfile.write(json.dumps({
+                    "error": "Unauthorized: Invalid or missing API key"
+                }).encode())
+                return
+
+        # Get request body
+        content_length = int(self.headers.get('Content-Length', 0))
+        post_data = self.rfile.read(content_length).decode('utf-8')
+        
+        try:
+            data = json.loads(post_data)
+            poll_id = data.get('pollId')
+            force = data.get('force', False)
+            
+            if not poll_id:
+                self.send_response(400)
+                self.send_header('Content-type', 'application/json')
+                self.end_headers()
+                self.wfile.write(json.dumps({
+                    "error": "Missing required parameter: pollId"
+                }).encode())
+                return
+            
+            logger.info(f"Triggering GAC update for poll: {poll_id}")
+            
+            # Run the GAC update for the specific poll
+            result = main(poll_id=poll_id, force=force)
+            
+            # Send success response
+            self.send_response(200)
+            self.send_header('Content-type', 'application/json')
+            self.end_headers()
+            self.wfile.write(json.dumps({
+                "success": True,
+                "message": f"GAC update triggered for poll: {poll_id}",
+                "result": result
+            }).encode())
+            
+        except json.JSONDecodeError:
+            self.send_response(400)
+            self.send_header('Content-type', 'application/json')
+            self.end_headers()
+            self.wfile.write(json.dumps({
+                "error": "Invalid JSON in request body"
+            }).encode())
+        except Exception as e:
+            logger.error(f"Error processing GAC update: {str(e)}")
+            self.send_response(500)
+            self.send_header('Content-type', 'application/json')
+            self.end_headers()
+            self.wfile.write(json.dumps({
+                "error": f"Internal server error: {str(e)}"
+            }).encode())
 
 def create_connection():
     # Reload DATABASE_URL at runtime
