+++ apps/consensus-service/package.json
@@ -13,7 +13,7 @@
     "gac:prod": "dotenv -e .env -- python api/update_gac_scores.py",
     "gac:prod:force": "dotenv -e .env -- python api/update_gac_scores.py --force",
     "gac:prod:force:dry": "dotenv -e .env -- python api/update_gac_scores.py --force --dry-run",
-    "serve:local": "dotenv -e .env.local -- python local_server.py",
+    "serve:local": "PYTHONPATH=$PYTHONPATH:. dotenv -e .env.local -- python local_server.py",
     "build": "echo 'Starting build process from package.json' && pip install -r api/requirements.txt --target ./python_packages && echo 'Requirements installed'",
     "test": "PYTHONPATH=$PYTHONPATH:. pytest tests/ -v",
     "test:install": "pip install -r api/requirements.txt && pnpm test"
